test: split react ui primitive tests
