+++ packages/react-ui/test/message.test.tsx
@@ -1,0 +1,170 @@
+import type { UIMessage } from "@anvia/react";
+import { render, screen } from "@testing-library/react";
+import { describe, expect, it } from "vitest";
+
+import { ChatProvider, Message, Thread } from "../src";
+import { createChatController } from "./helpers";
+
+describe("Message primitives", () => {
+  it("renders chat messages and all supported message part kinds", () => {
+    const messages: UIMessage[] = [
+      {
+        id: "assistant_1",
+        role: "assistant",
+        parts: [
+          { id: "text_1", type: "text", text: "Hello" },
+          { id: "reasoning_1", type: "reasoning", text: "Thinking" },
+          {
+            id: "tool_1",
+            type: "tool",
+            toolName: "search",
+            toolCallId: "call_1",
+            state: "output-available",
+            input: { query: "anvia" },
+            output: { ok: true },
+          },
+          { id: "data_1", type: "data", name: "result", data: { count: 2 } },
+          { id: "error_1", type: "error", error: { message: "Nope" } },
+        ],
+      },
+    ];
+
+    render(
+      <ChatProvider controller={createChatController({ messages })}>
+        <Thread.Root>
+          <Thread.Messages>
+            <Message.Root>
+              <Message.Content>
+                <Message.Parts />
+              </Message.Content>
+            </Message.Root>
+          </Thread.Messages>
+        </Thread.Root>
+      </ChatProvider>,
+    );
+
+    expect(screen.getByText("Hello")).toBeTruthy();
+    expect(screen.getByText("Thinking")).toBeTruthy();
+    expect(screen.getByText("search")).toBeTruthy();
+    expect(screen.getByText(/"query": "anvia"/)).toBeTruthy();
+    expect(screen.getByText(/"count": 2/)).toBeTruthy();
+    expect(screen.getByText("Nope")).toBeTruthy();
+  });
+
+  it("supports custom tool rendering with input and output in one component", () => {
+    const messages: UIMessage[] = [
+      {
+        id: "assistant_1",
+        role: "assistant",
+        parts: [
+          {
+            id: "tool_1",
+            type: "tool",
+            toolName: "get_order",
+            toolCallId: "call_1",
+            state: "output-available",
+            input: { id: "A-100" },
+            output: { status: "blocked" },
+          },
+        ],
+      },
+    ];
+
+    render(
+      <ChatProvider controller={createChatController({ messages })}>
+        <Thread.Root>
+          <Thread.Messages>
+            <Message.Root>
+              <Message.Content>
+                <Message.Parts>
+                  {(part) =>
+                    part.type === "tool" ? (
+                      <Message.Part>
+                        <Message.Tool>
+                          {(tool) => (
+                            <section data-testid="tool-card">
+                              <h2>{tool.toolName}</h2>
+                              <div>Input: {JSON.stringify(tool.input)}</div>
+                              <div>Output: {JSON.stringify(tool.output)}</div>
+                            </section>
+                          )}
+                        </Message.Tool>
+                      </Message.Part>
+                    ) : (
+                      <Message.Part />
+                    )
+                  }
+                </Message.Parts>
+              </Message.Content>
+            </Message.Root>
+          </Thread.Messages>
+        </Thread.Root>
+      </ChatProvider>,
+    );
+
+    expect(screen.getAllByTestId("tool-card")).toHaveLength(1);
+    expect(screen.getByText("get_order")).toBeTruthy();
+    expect(screen.getByText('Input: {"id":"A-100"}')).toBeTruthy();
+    expect(screen.getByText('Output: {"status":"blocked"}')).toBeTruthy();
+  });
+
+  it("filters message parts and gates tool rendering by state", () => {
+    const messages: UIMessage[] = [
+      {
+        id: "assistant_1",
+        role: "assistant",
+        parts: [
+          {
+            id: "tool_pending",
+            type: "tool",
+            toolName: "pending_lookup",
+            toolCallId: "call_pending",
+            state: "input-available",
+            input: { id: "A-100" },
+          },
+          {
+            id: "tool_done",
+            type: "tool",
+            toolName: "done_lookup",
+            toolCallId: "call_done",
+            state: "output-available",
+            input: { id: "B-200" },
+            output: { status: "shipped" },
+          },
+        ],
+      },
+    ];
+
+    render(
+      <ChatProvider controller={createChatController({ messages })}>
+        <Thread.Root>
+          <Thread.Messages>
+            <Message.Root>
+              <Message.Content>
+                <Message.Parts
+                  filter={(part) => part.type !== "tool" || part.state === "output-available"}
+                >
+                  {(part) =>
+                    part.type === "tool" ? (
+                      <Message.Part>
+                        <Message.Tool renderWhen="settled">
+                          {(tool) => <span data-testid="settled-tool">{tool.toolName}</span>}
+                        </Message.Tool>
+                      </Message.Part>
+                    ) : (
+                      <Message.Part />
+                    )
+                  }
+                </Message.Parts>
+              </Message.Content>
+            </Message.Root>
+          </Thread.Messages>
+        </Thread.Root>
+      </ChatProvider>,
+    );
+
+    expect(screen.queryByText("pending_lookup")).toBeNull();
+    expect(screen.getByText("done_lookup")).toBeTruthy();
+    expect(screen.getAllByTestId("settled-tool")).toHaveLength(1);
+  });
+});
